cinema-prediction/requirements.txt
# Add your Python dependencies here 
flask
pandas
plotly
numpy
scikit-learn
python-dotenv
